requirements.txt
# Core MCP dependencies
fastmcp>=2.0.0

# Neo4j dependencies
neo4j>=5.0.0

# Data processing
pydantic>=2.0.0

# Configuration and utilities
python-dotenv>=1.0.0
